requirements.txt
asgiref==3.8.1
cffi==1.17.1
Django==5.1.2
pillow==10.4.0
psycopg==3.2.3
pycparser==2.22
python-decouple==3.8
sqlparse==0.5.1
